minor improvements, custom exception

+++ src/main/java/com/example/pizdaa/service/PostServiceImpl.java
@@ -2,8 +2,10 @@ package com.example.pizdaa.service;
 
 import com.example.pizdaa.dto.PostDTO;
 import com.example.pizdaa.dto.mapping.PostMapping;
+import com.example.pizdaa.exception.PostNotFoundException;
 import com.example.pizdaa.model.PostModel;
 import com.example.pizdaa.repository.PostRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,7 +21,7 @@ public class PostServiceImpl implements PostService {
         this.postMapping = postMapping;
     }
 
-    // Получаем все посты в виде DTO
+
     @Override
     public List<PostDTO> getAllPosts() {
         return postRepository.findAll().stream()
@@ -27,15 +29,15 @@ public class PostServiceImpl implements PostService {
                 .toList();
     }
 
-    // Получаем один пост по id в виде DTO
+
     @Override
     public PostDTO getPostById(Integer id) {
         PostModel post = postRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Post not found with id " + id));
+                .orElseThrow(() -> new PostNotFoundException("Post not found with id " + id));
         return postMapping.toDTO(post);
     }
 
-    // Сохраняем новый пост, принимая DTO
+
     @Override
     public PostDTO savePost(PostDTO postDTO) {
         PostModel postModel = postMapping.toEntity(postDTO);
@@ -43,17 +45,19 @@ public class PostServiceImpl implements PostService {
         return postMapping.toDTO(savedPost);
     }
 
-    // Удаляем пост
     @Override
     public void deletePost(Integer id) {
+        if (!postRepository.existsById(id)) {
+            throw new PostNotFoundException("Post not found with id " + id);
+        }
         postRepository.deleteById(id);
     }
 
-    // Обновляем пост по id, принимая DTO
+    @Transactional
     @Override
     public PostDTO updatePost(Integer id, PostDTO postDTO) {
         PostModel existingPost = postRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Post not found with id " + id));
+                .orElseThrow(() -> new PostNotFoundException("Post not found with id " + id));
 
         if (postDTO.getTitle() != null) existingPost.setTitle(postDTO.getTitle());
         if (postDTO.getContent() != null) existingPost.setContent(postDTO.getContent());
